+++ service/src/main/java/org/ehrbase/dao/access/interfaces/I_CompositionAccess.java
@@ -484,4 +484,9 @@ public interface I_CompositionAccess extends I_SimpleCRUD {
     static boolean isDeleted(I_DomainAccess domainAccess, UUID versionedObjectId) {
         return CompositionAccess.isDeleted(domainAccess, versionedObjectId);
     }
+
+    /**
+     * Invoke physical deletion.
+     */
+    void adminDelete();
 }