+++ service/src/main/java/org/ehrbase/dao/access/jooq/EhrAccess.java
@@ -40,13 +40,16 @@ import org.ehrbase.dao.access.interfaces.*;
 import org.ehrbase.dao.access.jooq.party.PersistedPartyProxy;
 import org.ehrbase.dao.access.support.DataAccess;
 import org.ehrbase.dao.access.util.ContributionDef;
+import org.ehrbase.jooq.pg.Routines;
 import org.ehrbase.dao.access.util.TransactionTime;
 import org.ehrbase.jooq.pg.enums.ContributionDataType;
+import org.ehrbase.jooq.pg.tables.*;
 import org.ehrbase.jooq.pg.tables.records.*;
 import org.ehrbase.serialisation.dbencoding.RawJson;
 import org.ehrbase.service.RecordedDvCodedText;
 import org.ehrbase.service.RecordedDvText;
 import org.jooq.DSLContext;
+import org.jooq.Field;
 import org.jooq.Record;
 import org.jooq.Result;
 
@@ -742,5 +745,44 @@ public class EhrAccess extends DataAccess implements I_EhrAccess {
         status.setSubject(partySelf);
 
         return status;
+    }
+
+    @Override
+    public void adminDeleteEhr() {
+        AdminApiUtils adminApi = new AdminApiUtils(getContext());
+
+        // retrieve linked entities
+        Result<AdminGetLinkedCompositionsRecord> linkedCompositions = Routines.adminGetLinkedCompositions(getContext().configuration(), this.getId());
+        Result<AdminGetLinkedContributionsRecord> linkedContributions = Routines.adminGetLinkedContributions(getContext().configuration(), this.getId());
+
+        // remove linked directory folder
+        // TODO: get linked folders from "folder" attribute, once this RM 1.1.0 attribute is implemented
+        if (getDirectoryId() != null)
+            adminApi.deleteFolder(getDirectoryId(), false); // contribs will be deleted by EHR later
+
+        // handling of existing composition
+        linkedCompositions.forEach(compo -> adminApi.deleteComposition(compo.getComposition()));
+
+        // delete EHR itself
+        Result<AdminDeleteEhrRecord> adminDeleteEhr = Routines.adminDeleteEhr(getContext().configuration(), this.getId());
+        if (adminDeleteEhr.isEmpty()) {
+            throw new InternalServerException("Admin deletion of EHR failed! Unexpected result.");
+        }
+
+        // adminDeleteEhr returns all linked contributions, audits and statuses - go through and delete them
+        adminDeleteEhr.forEach(response -> {
+            UUID statusAudit = response.getStatusAudit();
+
+            // delete status audit
+            adminApi.deleteAudit(statusAudit, "Status", false);
+
+            // delete linked contributions
+            linkedContributions.forEach(contrib -> adminApi.deleteContribution(contrib.getContribution(), contrib.getAudit(), true));
+
+            // final cleanup of auxiliary objects
+            int res = getContext().selectQuery(new AdminDeleteEhrHistory().call(this.getId())).execute();
+            if (res != 1)
+                throw new InternalServerException("Admin deletion of EHR failed!");
+        });
     }
 }