+++ service/src/main/java/org/ehrbase/service/ContributionServiceImp.java
@@ -50,6 +50,7 @@ import org.jooq.DSLContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -326,5 +327,12 @@ public class ContributionServiceImp extends BaseService implements ContributionS
         DvText description = new DvText(auditDetailsAccess.getDescription());
 
         return new AuditDetails(systemId, committer, timeCommitted, changeType, description);
+    }
+
+    @PreAuthorize("hasRole('ADMIN')")
+    @Override
+    public void adminDelete(UUID contributionId) {
+        I_ContributionAccess contributionAccess = I_ContributionAccess.retrieveInstance(getDataAccess(), contributionId);
+        contributionAccess.adminDelete();
     }
 }